Switch to Terraform local backend (#44)
This PR changes the Terraform Zone codebase from a demo into an example.

The demo was too much configured to our need of running it as demo. It didn't make it easy for a user to download it and get started. It also used DNSimple internal settings are placeholders.

We have now a dedicated demo in a separate repository that is connected with Terraform Cloud and GitHub Actions. Therefore we can simplify this example.

The PR contains the following changes:

- Switched to the default local backend, so that the user can get started immediately
- Fixed an issue where the user could not run `terraform init` without setting up Terraform Cloud and updating the config (hence instructions could not be followed step by step)
- Removed all action workflows, as they are relevant only for the demo. The user may decide to move beyond a local Terraform setup, but that choice is very specific to user's needs. However, I included some instructions to where they can find more instructions.
- Changes to the default zone records to be more generic, and not focused on what we wanted to run as Demo
- Fixed token string format (it requires quotes)